+++ lib/yamload/loader.rb
@@ -1,5 +1,8 @@
 require 'yaml'
 require 'ice_nine'
+require 'aws-sdk-secretsmanager'
+require 'aws-sdk-ssm'
+
 require 'yamload/loading'
 require 'yamload/conversion'
 require 'yamload/defaults'