+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+ - [TT-7683] Add support for AWS SSM and AWS Secrets Manager
+
 ## 0.4.0 (2020-05-20)
 
  - [TT-7323] Bring all dependencies up to date